Module Pvec

Persistent vectors. Bitmapped trie + tail optimization. Branching factor 32.

module type T = sig ... end

include T

type 'a t

Vector holding elements of type 'a.

val length : 'a t -> int

length v returns the length of vector v.

val empty : unit -> 'a t

empty () returns the new length of vector of length 0.

val append : 'a -> 'a t -> 'a t

append x v appends x to the end of vector v and returns the updated vector.

val get : int -> 'a t -> 'a option

get i v reads the i-th element from vector v.

Returns None if index i is out of bounds.

val set : int -> 'a -> 'a t -> 'a t option

set i x v replaces the i-th element of vector v with x and returns the updated vector. If i = length v the element x is appended to v.

Returns None if index i is out of bounds.

val peek : 'a t -> 'a option

peek v returns the last element of vector v or None if v is empty.

val pop : 'a t -> ('a * 'a t) option

pop v removes the last element of vector v and returns the removed element together with the updated vector.

Returns None if v is empty.

val get_exn : int -> 'a t -> 'a

get_exn is similar to get but raises Not_found instead of returning None.

val set_exn : int -> 'a -> 'a t -> 'a t

set_exn is similar to set but raises Invalid_argument _ instead of returning None.

val to_seq : 'a t -> 'a Stdlib.Seq.t

to_seq v iterates over vector v front-to-back.

val rev_to_seq : 'a t -> 'a Stdlib.Seq.t

rev_to_seq v iterates over vector v back-to-front.

val of_seq : 'a Stdlib.Seq.t -> 'a t

of_seq s stores the elements of sequence s in a vector.

val to_list : 'a t -> 'a list

to_list v converts vector v to a list.

val of_list : 'a list -> 'a t

of_list v converts list l to a vector.

val init : int -> ( int -> 'a ) -> 'a t

init n f returns a vector of length n holding elements f(0), f(1), ... .

val iter : ( 'a -> unit ) -> 'a t -> unit

iter f v applies f to elements of v in order of the vector's index.

val rev_iter : ( 'a -> unit ) -> 'a t -> unit

iter f v applies f to elements of v in reverse order.

val debug_pp : Stdlib.Format.formatter -> 'a t -> unit

module Make (_ : sig ... end) : T

Create vectors with custom branching factor.